refactor!: enforce naming convention of packages with singular and plural
