+++ optree/ops.py
@@ -26,7 +26,7 @@ from collections import OrderedDict, defaultdict, deque
 from typing import TYPE_CHECKING, Any, Callable, ClassVar, Generic, overload
 
 import optree._C as _C
-from optree.accessor import PyTreeAccessor
+from optree.accessors import PyTreeAccessor
 from optree.typing import NamedTuple, T, is_namedtuple_instance, is_structseq_instance
 
 
@@ -34,7 +34,7 @@ if TYPE_CHECKING:
     import builtins
     from collections.abc import Collection, Iterable, Mapping
 
-    from optree.accessor import PyTreeEntry
+    from optree.accessors import PyTreeEntry
     from optree.typing import MetaData, PyTree, PyTreeKind, PyTreeSpec, S, U, UnflattenFunc
     from optree.typing import structseq as StructSequence  # noqa: N812
